Support rotation in new dialog based KeyChainActivity
Using showDialog, the framework manged the dialog for me. So on
rotation, it skiped onResume and right to onCreateDialog but then the
mCertificateAdapter was null. I considered showing the list first, and
filling in the content later, more like I do
TrustedCredentialsSettings. However, here the dialog content is
different if there are no entries.

To fix, I just don't use showDialog. Then onResume is called and I
recreate the dialog the same way in both cases.

Bug: 4967378
Change-Id: I306e3a32069dce4e00864be79b50e081869851d5